fix bugs EmployeService

+++ src/main/java/tn/esprit/spring/services/EmployeServiceImpl.java
@@ -3,6 +3,7 @@ package tn.esprit.spring.services;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -44,41 +45,55 @@ public class EmployeServiceImpl implements IEmployeService {
 
 
 	public void mettreAjourEmailByEmployeId(String email, int employeId) {
-		Employe employe = employeRepository.findById(employeId).get();
-		employe.setEmail(email);
-		employeRepository.save(employe);
+		Optional<Employe> employeOpt = employeRepository.findById(employeId);
+		if(employeOpt.isPresent()) {
+			Employe employe = employeOpt.get();
+			employe.setEmail(email);
+			employeRepository.save(employe);
+		}
+		
+
 
 	}
 
 	@Transactional	
 	public void affecterEmployeADepartement(int employeId, int depId) {
-		Departement depManagedEntity = deptRepoistory.findById(depId).get();
-		Employe employeManagedEntity = employeRepository.findById(employeId).get();
+		Optional<Employe> employeOpt = employeRepository.findById(employeId);
+		Optional<Departement> departementOpt = deptRepoistory.findById(depId);
+		
+		if(employeOpt.isPresent() && departementOpt.isPresent()) {
+			Departement depManagedEntity = departementOpt.get();
+			Employe employeManagedEntity = employeOpt.get();
 
-		if(depManagedEntity.getEmployes() == null){
+			if(depManagedEntity.getEmployes() == null){
 
-			List<Employe> employes = new ArrayList<>();
-			employes.add(employeManagedEntity);
-			depManagedEntity.setEmployes(employes);
-		}else{
+				List<Employe> employes = new ArrayList<>();
+				employes.add(employeManagedEntity);
+				depManagedEntity.setEmployes(employes);
+			}else{
 
-			depManagedEntity.getEmployes().add(employeManagedEntity);
+				depManagedEntity.getEmployes().add(employeManagedEntity);
+			}
+
+			// à ajouter? 
+			deptRepoistory.save(depManagedEntity); 			
 		}
 
-		// à ajouter? 
-		deptRepoistory.save(depManagedEntity); 
 
 	}
 	@Transactional
 	public void desaffecterEmployeDuDepartement(int employeId, int depId)
 	{
-		Departement dep = deptRepoistory.findById(depId).get();
+		Optional<Departement> departementOpt = deptRepoistory.findById(depId);
+		if(departementOpt.isPresent()) {
+			Departement dep = departementOpt.get();
 
-		int employeNb = dep.getEmployes().size();
-		for(int index = 0; index < employeNb; index++){
-			if(dep.getEmployes().get(index).getId() == employeId){
-				dep.getEmployes().remove(index);
-				break;//a revoir
+			int employeNb = dep.getEmployes().size();
+			for(int index = 0; index < employeNb; index++){
+				if(dep.getEmployes().get(index).getId() == employeId){
+					dep.getEmployes().remove(index);
+					break;//a revoir
+				}
 			}
 		}
 	} 
@@ -91,36 +106,49 @@ public class EmployeServiceImpl implements IEmployeService {
 	}
 
 	public void affecterContratAEmploye(int contratId, int employeId) {
-		Contrat contratManagedEntity = contratRepoistory.findById(contratId).get();
-		Employe employeManagedEntity = employeRepository.findById(employeId).get();
+		Optional<Employe> employeOpt = employeRepository.findById(employeId);
+		Optional<Contrat> contratOpt = contratRepoistory.findById(contratId);
+		
+		if(employeOpt.isPresent() && contratOpt.isPresent()) {
+			Contrat contratManagedEntity = contratOpt.get();
+			Employe employeManagedEntity = employeOpt.get();
 
-		contratManagedEntity.setEmploye(employeManagedEntity);
-		contratRepoistory.save(contratManagedEntity);
+			contratManagedEntity.setEmploye(employeManagedEntity);
+			contratRepoistory.save(contratManagedEntity);
+		}
 
 	}
 
 	public String getEmployePrenomById(int employeId) {
-		Employe employeManagedEntity = employeRepository.findById(employeId).get();
-		return employeManagedEntity.getPrenom();
+		Optional<Employe> employeOpt = employeRepository.findById(employeId);
+		if(employeOpt.isPresent()) {
+			Employe employeManagedEntity = employeOpt.get();
+			return employeManagedEntity.getPrenom();
+		}
+		return "";
 	}
 	 
 	public void deleteEmployeById(int employeId)
 	{
-		Employe employe = employeRepository.findById(employeId).get();
+		
+		Optional<Employe> employeOpt = employeRepository.findById(employeId);
+		if(employeOpt.isPresent()) {
+			Employe employe = employeOpt.get();
+			for(Departement dep : employe.getDepartements()){
+				dep.getEmployes().remove(employe);
+			}
 
-		//Desaffecter l'employe de tous les departements
-		//c'est le bout master qui permet de mettre a jour
-		//la table d'association
-		for(Departement dep : employe.getDepartements()){
-			dep.getEmployes().remove(employe);
+			employeRepository.delete(employe);
 		}
-
-		employeRepository.delete(employe);
 	}
 
 	public void deleteContratById(int contratId) {
-		Contrat contratManagedEntity = contratRepoistory.findById(contratId).get();
-		contratRepoistory.delete(contratManagedEntity);
+		Optional<Contrat> ContratOpt = contratRepoistory.findById(contratId);
+		if(ContratOpt.isPresent()) {
+			Contrat contratManagedEntity = ContratOpt.get();
+			contratRepoistory.delete(contratManagedEntity);			
+		}
+
 
 	}
